requirements.txt
streamlit
pytesseract
Pillow
openai
replicate
